requirements.txt
pandasql==0.7.3
streamlit
statsmodels==0.13.2
